everything is working now: deb, rpm, tgz, wheel

+++ cpp/scripts/_build_musl_wheel.sh
@@ -1,11 +1,25 @@
-#!/bin/bash
+#!/bin/bash -x
 
 source $(dirname $0)/common.sh
 
 # This script is used to build a wheel for the current package.
 
-TMP_DEST_DIR="tmp_dist"
-DEST_DIR="dist"
+export CMAKE_PROJECT_VERSION=${PROJECT_VERSION}
+$(dirname $0)/_build_atomdb_lib_with_musl.sh
+
+# Check if the RPM files are present
+if [ ! -e ${LIB_DIST_DIR}/${LIB_NAME_PATTERN}*.tar.gz 2>&1 >/dev/null ]; then
+    echo "Error: No tar.gz files found in ${LIB_DIST_DIR}"
+    exit 1
+fi
+
+# Install the RPM file
+tar zvxf ${LIB_DIST_DIR}/${LIB_NAME_PATTERN}*.tar.gz --strip-components=1 -C /
+[ $? -ne 0 ] && exit 1
+
+NANOBIND_ROOT=$(realpath "${PWD}/nanobind")
+TMP_DEST_DIR="${NANOBIND_ROOT}/tmp_dist"
+DEST_DIR="${NANOBIND_ROOT}/dist"
 
 # Remove existing/old wheel files
 rm -rf ${TMP_DEST_DIR}
@@ -16,11 +30,12 @@ mkdir -p ${TMP_DEST_DIR} ${DEST_DIR}
 [ $? -ne 0 ] && exit 1
 
 # Build wheel
+cd ${NANOBIND_ROOT}
 ${PYTHON_EXECUTABLE} -m pip wheel . --wheel-dir ${TMP_DEST_DIR}
 [ $? -ne 0 ] && exit 1
 
 # Repair wheel - tag as manylinux
-find ${TMP_DEST_DIR} -type f -name "*abi3-linux_x86_64.whl" \
+find ${TMP_DEST_DIR} -type f -name "*.whl" \
   -exec ${PYTHON_EXECUTABLE} -m auditwheel repair {} -w ${DEST_DIR} \;
 [ $? -ne 0 ] && exit 1
